backend/requirements.txt
fastapi
uvicorn
python-multipart
pymupdf
easyocr
python-docx

